// src/app/services/projects/models/project-stats-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProjectStatsDto {
  activeSprints?: number;
  averageStoryPointsPerTask?: number;
  averageTasksPerSprint?: number;
  blockedTasks?: number;
  completedSprints?: number;
  completedTasks?: number;
  completionRate?: number;
  inProgressTasks?: number;
  projectId?: string;
  projectName?: string;
  tasksByPriority?: {
[key: string]: number;
};
  tasksByType?: {
[key: string]: number;
};
  todoTasks?: number;
  totalMembers?: number;
  totalSprints?: number;
  totalTasks?: number;
}